backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-fido2==1.1.2
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.0.7
pyjwt==2.9.0
